Agentic-Screensaver/AgenticScreensaver/requirements.txt
pygame
psutil
google-generativeai
numpy
